requirements.txt
python-docx==1.2.0
sentence-transformers==4.1.0
faiss-cpu==1.11.0
oci==2.154.2
fastapi==0.115.13
uvicorn==0.34.3
